密钥与统计cast错误

+++ src/main/java/com/equator/course/service/cache/RequestTimesCache.java
@@ -19,7 +19,7 @@ public class RequestTimesCache {
         redisTemplate.opsForValue().increment(CacheKeyEnum.REQUEST_TIMES_NEVER.getValue());
     }
 
-    public long get(){
-        return (long) redisTemplate.opsForValue().get(CacheKeyEnum.REQUEST_TIMES_NEVER.getValue());
+    public Integer get() {
+        return (Integer) redisTemplate.opsForValue().get(CacheKeyEnum.REQUEST_TIMES_NEVER.getValue());
     }
 }
